PETSC

Documentation for PETSC.

PETSC.PETSC_DECIDE

PETSC.PETSC_DEFAULT

PETSC.PETSC_DETERMINE

PETSC.InsertMode

PETSC.KSP

PETSC.KSPType

PETSC.Mat

PETSC.MatAssemblyType

PETSC.MatDuplicateOption

PETSC.MatInfo

PETSC.MatInfoType

PETSC.MatNullSpace

PETSC.MatReuse

PETSC.MatStructure

PETSC.MatType

PETSC.NormType

PETSC.PC

PETSC.PCType

PETSC.PetscBool

PETSC.PetscDataType

PETSC.PetscError

PETSC.PetscErrorCode

PETSC.PetscLogDouble

PETSC.PetscViewer

PETSC.SNES

PETSC.Vec

PETSC.VecOption

PETSC.KSPCreate

PETSC.KSPDestroy

PETSC.KSPGetIterationNumber

PETSC.KSPGetPC

PETSC.KSPSetFromOptions

PETSC.KSPSetInitialGuessNonzero

PETSC.KSPSetOperators

PETSC.KSPSetOptionsPrefix

PETSC.KSPSetTolerances

PETSC.KSPSetType

PETSC.KSPSetUp

PETSC.KSPSolve

PETSC.KSPSolveTranspose

PETSC.KSPView

PETSC.MatAssemblyBegin

PETSC.MatAssemblyEnd

PETSC.MatConvert

PETSC.MatCopy

PETSC.MatCreateAIJ

PETSC.MatCreateMPIAIJWithArrays

PETSC.MatCreateSeqAIJ

PETSC.MatCreateSeqAIJWithArrays

PETSC.MatDestroy

PETSC.MatEqual

PETSC.MatGetInfo

PETSC.MatGetSize

PETSC.MatGetValues

PETSC.MatMult

PETSC.MatMultAdd

PETSC.MatMumpsSetCntl

PETSC.MatMumpsSetIcntl

PETSC.MatNullSpaceCreate

PETSC.MatNullSpaceCreateRigidBody

PETSC.MatNullSpaceDestroy

PETSC.MatScale

PETSC.MatSetBlockSize

PETSC.MatSetNearNullSpace

PETSC.MatSetValues

PETSC.MatView

PETSC.MatZeroEntries

PETSC.PCFactorGetMatrix

PETSC.PCFactorSetMatSolverType

PETSC.PCFactorSetUpMatSolverType

PETSC.PCSetType

PETSC.PCView

PETSC.PETSC_VIEWER_DRAW_

PETSC.PETSC_VIEWER_STDOUT_

PETSC.PetscDataTypeFromString

PETSC.PetscDataTypeGetSize

PETSC.PetscFinalize

PETSC.PetscFinalized

PETSC.PetscInitializeNoArguments

PETSC.PetscInitializeNoPointers

PETSC.PetscInitialized

PETSC.PetscObjectRegisterDestroy

PETSC.PetscObjectRegisterDestroyAll

PETSC.SNESCreate

PETSC.SNESDestroy

PETSC.SNESGetIterationNumber

PETSC.SNESGetKSP

PETSC.SNESGetLinearSolveFailures

PETSC.SNESGetLinearSolveIterations

PETSC.SNESGetNumberFunctionEvals

PETSC.SNESSetCountersReset

PETSC.SNESSetFromOptions

PETSC.SNESSetFunction

PETSC.SNESSetJacobian

PETSC.SNESSetType

PETSC.SNESSolve

PETSC.SNESView

PETSC.VecAXPBY

PETSC.VecAXPY

PETSC.VecAYPX

PETSC.VecAssemblyBegin

PETSC.VecAssemblyEnd

PETSC.VecCopy

PETSC.VecCreateGhost

PETSC.VecCreateGhostWithArray

PETSC.VecCreateMPI

PETSC.VecCreateSeq

PETSC.VecCreateSeqWithArray

PETSC.VecDestroy

PETSC.VecDuplicate

PETSC.VecGetArray

PETSC.VecGetArrayRead

PETSC.VecGetArrayWrite

PETSC.VecGetLocalSize

PETSC.VecGetSize

PETSC.VecGetValues

PETSC.VecGhostGetLocalForm

PETSC.VecGhostRestoreLocalForm

PETSC.VecNorm

PETSC.VecPlaceArray

PETSC.VecResetArray

PETSC.VecRestoreArray

PETSC.VecRestoreArrayRead

PETSC.VecRestoreArrayWrite

PETSC.VecScale

PETSC.VecSet

PETSC.VecSetOption

PETSC.VecSetValues

PETSC.VecView

PETSC.@PETSC_VIEWER_DRAW_SELF

PETSC.@PETSC_VIEWER_DRAW_WORLD

PETSC.@PETSC_VIEWER_STDOUT_SELF

PETSC.@PETSC_VIEWER_STDOUT_WORLD

PETSC.@check_error_code

PETSC.PETSC_DECIDE — Constant

Julia constant storing the PETSC_DECIDE value.

See PETSc manual.

PETSC.PETSC_DEFAULT — Constant

Julia constant storing the PETSC_DEFAULT value.

See PETSc manual.

PETSC.PETSC_DETERMINE — Constant

Julia constant storing the PETSC_DETERMINE value.

See PETSc manual.

PETSC.InsertMode — Type

Julia alias for the InsertMode C enum.

See PETSc manual.

PETSC.KSP — Type

Julia alias for the KSP C type.

See PETSc manual.

PETSC.KSPType — Type

Julia alias for KSPType C type.

See PETSc manual.

PETSC.Mat — Type

Julia alias for the Mat C type.

See PETSc manual.

PETSC.MatAssemblyType — Type

Julia alias for the MatAssemblyType C enum.

See PETSc manual.

PETSC.MatDuplicateOption — Type

Julia alias for the MatDuplicateOption C enum.

See PETSc manual.

PETSC.MatInfo — Type

Julia alias for the MatInfo C struct.

See PETSc manual.

PETSC.MatInfoType — Type

Julia alias for the MatInfoType C enum.

See PETSc manual.

PETSC.MatNullSpace — Type

Julia alias for the MatNullSpace C type.

See PETSc manual.

PETSC.MatReuse — Type

Julia alias for the MatReuse C enum.

See PETSc manual.

PETSC.MatStructure — Type

Julia alias for the MatStructure C enum.

See PETSc manual.

PETSC.MatType — Type

Julia alias for MatType C type.

See PETSc manual.

PETSC.NormType — Type

Julia alias for the NormType C enum.

See PETSc manual.

PETSC.PC — Type

Julia alias for the PC C type.

See PETSc manual.

PETSC.PCType — Type

Julia alias for PCType C type.

See PETSc manual.

PETSC.PetscBool — Type

Julia alias to PetscBool C enum.

See PETSc manual.

PETSC.PetscDataType — Type

Julia alias to PetscDataType C enum.

See PETSc manual.

PETSC.PetscError — Type

struct PetscError <: Exception
  code::PetscErrorCode
end

Custom Exception thrown by @check_error_code.

PETSC.PetscErrorCode — Type

Julia alias to PetscErrorCode C type.

See PETSc manual.

PETSC.PetscLogDouble — Type

Julia alias to PetscLogDouble C type.

See PETSc manual.

PETSC.PetscViewer — Type

Julia alias for PetscViewer C type.

See PETSc manual.

PETSC.SNES — Type

Julia alias for the SNES C type.

See PETSc manual.

PETSC.Vec — Type

Julia alias for the Vec C type.

See PETSc manual.

PETSC.VecOption — Type

Julia alias for the VecOption C enum.

See PETSc manual.

PETSC.KSPCreate — Method

KSPCreate(comm, inksp)

See PETSc manual.

PETSC.KSPDestroy — Method

KSPDestroy(ksp)

See PETSc manual.

PETSC.KSPGetIterationNumber — Method

KSPGetIterationNumber(ksp, its)

See PETSc manual.

PETSC.KSPGetPC — Method

KSPGetPC(ksp, pc)

See PETSc manual.

PETSC.KSPSetFromOptions — Method

KSPSetFromOptions(ksp)

See PETSc manual.

PETSC.KSPSetInitialGuessNonzero — Method

KSPSetInitialGuessNonzero(ksp, flg)

See PETSc manual.

PETSC.KSPSetOperators — Method

KSPSetOperators(ksp, Amat, Pmat)

See PETSc manual.

PETSC.KSPSetOptionsPrefix — Method

KSPSetOptionsPrefix(ksp, prefix)

See PETSc manual.

PETSC.KSPSetTolerances — Method

KSPSetTolerances(ksp, rtol, abstol, dtol, maxits)

See PETSc manual.

PETSC.KSPSetType — Method

KSPSetType(ksp, typ)

See PETSc manual.

PETSC.KSPSetUp — Method

KSPSetUp(ksp)

See PETSc manual.

PETSC.KSPSolve — Method

KSPSolve(ksp, b, x)

See PETSc manual.

PETSC.KSPSolveTranspose — Method

KSPSolveTranspose(ksp, b, x)

See PETSc manual.

PETSC.KSPView — Method

KSPView(ksp, viewer)

See PETSc manual.

PETSC.MatAssemblyBegin — Method

MatAssemblyBegin(mat, typ)

See PETSc manual.

PETSC.MatAssemblyEnd — Method

MatAssemblyEnd(mat, typ)

See PETSc manual.

PETSC.MatConvert — Method

MatConvert(mat, newtype, reuse, M)

See PETSc manual.

PETSC.MatCopy — Method

MatCopy(A, B, str)

See PETSc manual.

PETSC.MatCreateAIJ — Method

MatCreateAIJ(comm, m, n, M, N, d_nz, d_nnz, o_nz, o_nnz, mat)

See PETSc manual.

PETSC.MatCreateMPIAIJWithArrays — Method

MatCreateMPIAIJWithArrays(comm, m, n, M, N, i, j, a, mat)

See PETSc manual.

PETSC.MatCreateSeqAIJ — Method

MatCreateSeqAIJ(comm, m, n, nz, nnz, mat)

See PETSc manual.

PETSC.MatCreateSeqAIJWithArrays — Method

MatCreateSeqAIJWithArrays(comm, m, n, i, j, a, mat)

See PETSc manual.

PETSC.MatDestroy — Method

MatDestroy(A)

See PETSc manual.

PETSC.MatEqual — Method

MatEqual(A, B, flg)

See PETSc manual.

PETSC.MatGetInfo — Method

MatGetInfo(mat, flag, info)

See PETSc manual.

PETSC.MatGetSize — Method

MatGetSize(mat, m, n)

See PETSc manual.

PETSC.MatGetValues — Method

MatGetValues(mat, m, idxm, n, idxn, v)

See PETSc manual.

PETSC.MatMult — Method

MatMult(mat, x, y)

See PETSc manual.

PETSC.MatMultAdd — Method

MatMultAdd(mat, v1, v2, v3)

See PETSc manual.

PETSC.MatMumpsSetCntl — Method

MatMumpsSetCntl(mat, icntl, val)

See PETSc manual.

PETSC.MatMumpsSetIcntl — Method

MatMumpsSetIcntl(mat, icntl, val)

See PETSc manual.

PETSC.MatNullSpaceCreate — Method

MatNullSpaceCreate(comm, has_cnst, n, vecs, sp)

See PETSc manual.

PETSC.MatNullSpaceCreateRigidBody — Method

MatNullSpaceCreateRigidBody(coords, sp)

See PETSc manual.

PETSC.MatNullSpaceDestroy — Method

MatNullSpaceDestroy(ns)

See PETSc manual.

PETSC.MatScale — Method

MatScale(mat, alpha)

See PETSc manual.

PETSC.MatSetBlockSize — Method

MatSetBlockSize(mat, bs)

See PETSc manual.

PETSC.MatSetNearNullSpace — Method

MatSetNearNullSpace(mat, nullsp)

See PETSc manual.

PETSC.MatSetValues — Method

MatSetValues(mat, m, idxm, n, idxn, v, addv)

See PETSc manual.

PETSC.MatView — Method

MatView(mat, viewer)

See PETSc manual.

PETSC.MatZeroEntries — Method

MatZeroEntries(mat)

See PETSc manual.

PETSC.PCFactorGetMatrix — Method

PCFactorGetMatrix(ksp, mat)

See PETSc manual.

PETSC.PCFactorSetMatSolverType — Method

PCFactorSetMatSolverType(pc, typ)

See PETSc manual.

PETSC.PCFactorSetUpMatSolverType — Method

PCFactorSetUpMatSolverType(pc)

See PETSc manual.

PETSC.PCSetType — Method

PCSetType(pc, typ)

See PETSc manual.

PETSC.PCView — Method

PCView(pc, viewer)

See PETSc manual.

PETSC.PETSC_VIEWER_DRAW_ — Method

PETSC_VIEWER_DRAW_(comm)

See PETSc manual.

PETSC.PETSC_VIEWER_STDOUT_ — Method

PETSC_VIEWER_STDOUT_(comm)

See PETSc manual.

PETSC.PetscDataTypeFromString — Method

PetscDataTypeFromString(name,ptype,found)

See PETSc manual.

PETSC.PetscDataTypeGetSize — Method

PetscDataTypeGetSize(ptype,size)

See PETSc manual.

PETSC.PetscFinalize — Method

PetscFinalize()

See PETSc manual.

PETSC.PetscFinalized — Method

PetscFinalized(flag)

See PETSc manual.

PETSC.PetscInitializeNoArguments — Method

PetscInitializeNoArguments()

See PETSc manual.

PETSC.PetscInitializeNoPointers — Method

PetscInitializeNoPointers(argc, args, filename, help)

PETSC.PetscInitialized — Method

PetscInitialized(flag)

See PETSc manual.

PETSC.PetscObjectRegisterDestroy — Method

PetscObjectRegisterDestroy(obj)

See PETSc manual.

PETSC.PetscObjectRegisterDestroyAll — Method

PetscObjectRegisterDestroyAll()

See PETSc manual.

PETSC.SNESCreate — Method

SNESCreate(comm, snes)

See PETSc manual.

PETSC.SNESDestroy — Method

SNESDestroy(snes)

See PETSc manual.

PETSC.SNESGetIterationNumber — Method

SNESGetIterationNumber(snes, iter)

See PETSc manual.

PETSC.SNESGetKSP — Method

SNESGetKSP(snes, ksp)

See PETSc manual.

PETSC.SNESGetLinearSolveFailures — Method

SNESGetLinearSolveFailures(snes, nfails)

See PETSc manual.

PETSC.SNESGetLinearSolveIterations — Method

SNESGetLinearSolveIterations(snes, iter)

See PETSc manual.

PETSC.SNESGetNumberFunctionEvals — Method

SNESGetNumberFunctionEvals(snes, nfuncs)

See PETSc manual.

PETSC.SNESSetCountersReset — Method

SNESSetCountersReset(snes, reset)

See PETSc manual.

PETSC.SNESSetFromOptions — Method

SNESSetFromOptions(snes)

See PETSc manual.

PETSC.SNESSetFunction — Method

SNESSetFunction(snes, vec, fptr, ctx)

See PETSc manual.

PETSC.SNESSetJacobian — Method

SNESSetJacobian(snes, A, P, jacptr, ctx)

See PETSc manual.

PETSC.SNESSetType — Method

SNESSetType(snes, type)

See PETSc manual.

PETSC.SNESSolve — Method

SNESSolve(snes, b, x)

See PETSc manual.

PETSC.SNESView — Method

SNESView(snes, viewer)

See PETSc manual.

PETSC.VecAXPBY — Method

VecAXPBY(y, alpha, beta, x)

See PETSc manual.

PETSC.VecAXPY — Method

VecAXPY(y, alpha, x)

See PETSc manual.

PETSC.VecAYPX — Method

VecAYPX(y, beta, x)

See PETSc manual.

PETSC.VecAssemblyBegin — Method

VecAssemblyBegin(vec)

See PETSc manual.

PETSC.VecAssemblyEnd — Method

VecAssemblyEnd(vec)

See PETSc manual.

PETSC.VecCopy — Method

VecCopy(x, y)

See PETSc manual.

PETSC.VecCreateGhost — Method

VecCreateGhost(comm, n, N, nghost, ghosts, vv)

See PETSc manual.

PETSC.VecCreateGhostWithArray — Method

VecCreateGhostWithArray(comm, n, N, nghost, ghosts, array, vv)

See PETSc manual.

PETSC.VecCreateMPI — Method

VecCreateMPI(comm, n, N, v)

See PETSc manual.

PETSC.VecCreateSeq — Method

VecCreateSeq(comm, n, vec)

See PETSc manual.

PETSC.VecCreateSeqWithArray — Method

VecCreateSeqWithArray(comm, bs, n, array, vec)

See PETSc manual.

PETSC.VecDestroy — Method

VecDestroy(vec)

See PETSc manual.

PETSC.VecDuplicate — Method

VecDuplicate(v, newv)

See PETSc manual.

PETSC.VecGetArray — Method

VecGetArray(x, a)

See PETSc manual.

PETSC.VecGetArrayRead — Method

VecGetArrayRead(x, a)

See PETSc manual.

PETSC.VecGetArrayWrite — Method

VecGetArrayWrite(x, a)

See PETSc manual.

PETSC.VecGetLocalSize — Method

VecGetLocalSize(vec, n)

See PETSc manual.

PETSC.VecGetSize — Method

VecGetSize(vec, n)

See PETSc manual.

PETSC.VecGetValues — Method

VecGetValues(x, ni, ix, y)

See PETSc manual.

PETSC.VecGhostGetLocalForm — Method

VecGhostGetLocalForm(g, l)

See PETSc manual.

PETSC.VecGhostRestoreLocalForm — Method

VecGhostRestoreLocalForm(g, l)

See PETSc manual.

PETSC.VecNorm — Method

VecNorm(x, typ, val)

See PETSc manual.

PETSC.VecPlaceArray — Method

VecPlaceArray(vec, array)

See PETSc manual.

PETSC.VecResetArray — Method

VecResetArray(vec)

See PETSc manual.

PETSC.VecRestoreArray — Method

VecRestoreArray(x, a)

See PETSc manual.

PETSC.VecRestoreArrayRead — Method

VecRestoreArrayRead(x, a)

See PETSc manual.

PETSC.VecRestoreArrayWrite — Method

VecRestoreArrayWrite(x, a)

See PETSc manual.

PETSC.VecScale — Method

VecScale(x, alpha)

See PETSc manual.

PETSC.VecSet — Method

VecSet(x, alpha)

See PETSc manual.

PETSC.VecSetOption — Method

VecSetOption(x, op, flg)

See PETSc manual.

PETSC.VecSetValues — Method

VecSetValues(x, ni, ix, y, iora)

See PETSc manual.

PETSC.VecView — Method

VecView(vec, viewer)

See PETSc manual.

PETSC.@PETSC_VIEWER_DRAW_SELF — Macro

@PETSC_VIEWER_DRAW_SELF

See PETSc manual.

PETSC.@PETSC_VIEWER_DRAW_WORLD — Macro

@PETSC_VIEWER_DRAW_WORLD

See PETSc manual.

PETSC.@PETSC_VIEWER_STDOUT_SELF — Macro

@PETSC_VIEWER_STDOUT_SELF

See PETSc manual.

PETSC.@PETSC_VIEWER_STDOUT_WORLD — Macro

@PETSC_VIEWER_STDOUT_WORLD

See PETSc manual.

PETSC.@check_error_code — Macro

@check_error_code expr

Check if expr returns an error code equal to zero(PetscErrorCode). If not, throw an instance of PetscError.