Advanced API

PETSC.PETSC_DECIDE — Constant

Julia constant storing the PETSC_DECIDE value.

See PETSc manual.

PETSC.PETSC_DEFAULT — Constant

Julia constant storing the PETSC_DEFAULT value.

See PETSc manual.

PETSC.PETSC_DETERMINE — Constant

Julia constant storing the PETSC_DETERMINE value.

See PETSc manual.

PETSC.InsertMode — Type

Julia alias for the InsertMode C enum.

See PETSc manual.

PETSC.KSP — Type

Julia alias for the KSP C type.

See PETSc manual.

PETSC.KSPType — Type

Julia alias for KSPType C type.

See PETSc manual.

PETSC.Mat — Type

Julia alias for the Mat C type.

See PETSc manual.

PETSC.MatAssemblyType — Type

Julia alias for the MatAssemblyType C enum.

See PETSc manual.

PETSC.MatDuplicateOption — Type

Julia alias for the MatDuplicateOption C enum.

See PETSc manual.

PETSC.MatInfo — Type

Julia alias for the MatInfo C struct.

See PETSc manual.

PETSC.MatInfoType — Type

Julia alias for the MatInfoType C enum.

See PETSc manual.

PETSC.MatNullSpace — Type

Julia alias for the MatNullSpace C type.

See PETSc manual.

PETSC.MatReuse — Type

Julia alias for the MatReuse C enum.

See PETSc manual.

PETSC.MatStructure — Type

Julia alias for the MatStructure C enum.

See PETSc manual.

PETSC.MatType — Type

Julia alias for MatType C type.

See PETSc manual.

PETSC.NormType — Type

Julia alias for the NormType C enum.

See PETSc manual.

PETSC.PC — Type

Julia alias for the PC C type.

See PETSc manual.

PETSC.PCType — Type

Julia alias for PCType C type.

See PETSc manual.

PETSC.PetscBool — Type

Julia alias to PetscBool C enum.

See PETSc manual.

PETSC.PetscDataType — Type

Julia alias to PetscDataType C enum.

See PETSc manual.

PETSC.PetscError — Type

struct PetscError <: Exception
  code::PetscErrorCode
end

Custom Exception thrown by @check_error_code.

PETSC.PetscErrorCode — Type

Julia alias to PetscErrorCode C type.

See PETSc manual.

PETSC.PetscLogDouble — Type

Julia alias to PetscLogDouble C type.

See PETSc manual.

PETSC.PetscViewer — Type

Julia alias for PetscViewer C type.

See PETSc manual.

PETSC.SNES — Type

Julia alias for the SNES C type.

See PETSc manual.

PETSC.Vec — Type

Julia alias for the Vec C type.

See PETSc manual.

PETSC.VecOption — Type

Julia alias for the VecOption C enum.

See PETSc manual.

PETSC.KSPCreate — Method

PETSC.KSPCreate(comm, inksp)

See PETSc manual.

PETSC.KSPDestroy — Method

PETSC.KSPDestroy(ksp)

See PETSc manual.

PETSC.KSPGetIterationNumber — Method

PETSC.KSPGetIterationNumber(ksp, its)

See PETSc manual.

PETSC.KSPGetPC — Method

PETSC.KSPGetPC(ksp, pc)

See PETSc manual.

PETSC.KSPSetFromOptions — Method

PETSC.KSPSetFromOptions(ksp)

See PETSc manual.

PETSC.KSPSetInitialGuessNonzero — Method

PETSC.KSPSetInitialGuessNonzero(ksp, flg)

See PETSc manual.

PETSC.KSPSetOperators — Method

PETSC.KSPSetOperators(ksp, Amat, Pmat)

See PETSc manual.

PETSC.KSPSetOptionsPrefix — Method

PETSC.KSPSetOptionsPrefix(ksp, prefix)

See PETSc manual.

PETSC.KSPSetTolerances — Method

PETSC.KSPSetTolerances(ksp, rtol, abstol, dtol, maxits)

See PETSc manual.

PETSC.KSPSetType — Method

PETSC.KSPSetType(ksp, typ)

See PETSc manual.

PETSC.KSPSetUp — Method

PETSC.KSPSetUp(ksp)

See PETSc manual.

PETSC.KSPSolve — Method

PETSC.KSPSolve(ksp, b, x)

See PETSc manual.

PETSC.KSPSolveTranspose — Method

PETSC.KSPSolveTranspose(ksp, b, x)

See PETSc manual.

PETSC.KSPView — Method

PETSC.KSPView(ksp, viewer)

See PETSc manual.

PETSC.MatAssemblyBegin — Method

PETSC.MatAssemblyBegin(mat, typ)

See PETSc manual.

PETSC.MatAssemblyEnd — Method

PETSC.MatAssemblyEnd(mat, typ)

See PETSc manual.

PETSC.MatConvert — Method

PETSC.MatConvert(mat, newtype, reuse, M)

See PETSc manual.

PETSC.MatCopy — Method

PETSC.MatCopy(A, B, str)

See PETSc manual.

PETSC.MatCreateAIJ — Method

PETSC.MatCreateAIJ(comm, m, n, M, N, d_nz, d_nnz, o_nz, o_nnz, mat)

See PETSc manual.

PETSC.MatCreateMPIAIJWithArrays — Method

PETSC.MatCreateMPIAIJWithArrays(comm, m, n, M, N, i, j, a, mat)

See PETSc manual.

PETSC.MatCreateMPIAIJWithSplitArrays — Method

PETSC.MatCreateMPIAIJWithSplitArrays(comm, m, n, M, N, i, j, a, oi, oj, oa, mat)

See PETSc manual.

PETSC.MatCreateSeqAIJ — Method

PETSC.MatCreateSeqAIJ(comm, m, n, nz, nnz, mat)

See PETSc manual.

PETSC.MatCreateSeqAIJWithArrays — Method

PETSC.MatCreateSeqAIJWithArrays(comm, m, n, i, j, a, mat)

See PETSc manual.

PETSC.MatDestroy — Method

PETSC.MatDestroy(A)

See PETSc manual.

PETSC.MatEqual — Method

PETSC.MatEqual(A, B, flg)

See PETSc manual.

PETSC.MatGetInfo — Method

PETSC.MatGetInfo(mat, flag, info)

See PETSc manual.

PETSC.MatGetSize — Method

PETSC.MatGetSize(mat, m, n)

See PETSc manual.

PETSC.MatGetValues — Method

PETSC.MatGetValues(mat, m, idxm, n, idxn, v)

See PETSc manual.

PETSC.MatMult — Method

PETSC.MatMult(mat, x, y)

See PETSc manual.

PETSC.MatMultAdd — Method

PETSC.MatMultAdd(mat, v1, v2, v3)

See PETSc manual.

PETSC.MatMumpsSetCntl — Method

PETSC.MatMumpsSetCntl(mat, icntl, val)

See PETSc manual.

PETSC.MatMumpsSetIcntl — Method

PETSC.MatMumpsSetIcntl(mat, icntl, val)

See PETSc manual.

PETSC.MatNullSpaceCreate — Method

PETSC.MatNullSpaceCreate(comm, has_cnst, n, vecs, sp)

See PETSc manual.

PETSC.MatNullSpaceCreateRigidBody — Method

PETSC.MatNullSpaceCreateRigidBody(coords, sp)

See PETSc manual.

PETSC.MatNullSpaceDestroy — Method

PETSC.MatNullSpaceDestroy(ns)

See PETSc manual.

PETSC.MatScale — Method

PETSC.MatScale(mat, alpha)

See PETSc manual.

PETSC.MatSetBlockSize — Method

PETSC.MatSetBlockSize(mat, bs)

See PETSc manual.

PETSC.MatSetNearNullSpace — Method

PETSC.MatSetNearNullSpace(mat, nullsp)

See PETSc manual.

PETSC.MatSetValues — Method

PETSC.MatSetValues(mat, m, idxm, n, idxn, v, addv)

See PETSc manual.

PETSC.MatView — Method

PETSC.MatView(mat, viewer)

See PETSc manual.

PETSC.MatZeroEntries — Method

PETSC.MatZeroEntries(mat)

See PETSc manual.

PETSC.PCFactorGetMatrix — Method

PETSC.PCFactorGetMatrix(ksp, mat)

See PETSc manual.

PETSC.PCFactorSetMatSolverType — Method

PETSC.PCFactorSetMatSolverType(pc, typ)

See PETSc manual.

PETSC.PCFactorSetUpMatSolverType — Method

PETSC.PCFactorSetUpMatSolverType(pc)

See PETSc manual.

PETSC.PCSetType — Method

PETSC.PCSetType(pc, typ)

See PETSc manual.

PETSC.PCView — Method

PETSC.PCView(pc, viewer)

See PETSc manual.

PETSC.PETSC_VIEWER_DRAW_ — Method

PETSC.PETSC_VIEWER_DRAW_(comm)

See PETSc manual.

PETSC.PETSC_VIEWER_STDOUT_ — Method

PETSC.PETSC_VIEWER_STDOUT_(comm)

See PETSc manual.

PETSC.PetscDataTypeFromString — Method

PetscDataTypeFromString(name,ptype,found)

See PETSc manual.

PETSC.PetscDataTypeGetSize — Method

PetscDataTypeGetSize(ptype,size)

See PETSc manual.

PETSC.PetscFinalize — Method

PETSC.PetscFinalize()

See PETSc manual.

PETSC.PetscFinalized — Method

PETSC.PetscFinalized(flag)

See PETSc manual.

PETSC.PetscInitializeNoArguments — Method

PETSC.PetscInitializeNoArguments()

See PETSc manual.

PETSC.PetscInitializeNoPointers — Method

PETSC.PetscInitializeNoPointers(argc, args, filename, help)

PETSC.PetscInitialized — Method

PETSC.PetscInitialized(flag)

See PETSc manual.

PETSC.PetscObjectRegisterDestroy — Method

PETSC.PetscObjectRegisterDestroy(obj)

See PETSc manual.

PETSC.PetscObjectRegisterDestroyAll — Method

PETSC.PetscObjectRegisterDestroyAll()

See PETSc manual.

PETSC.SNESCreate — Method

PETSC.SNESCreate(comm, snes)

See PETSc manual.

PETSC.SNESDestroy — Method

PETSC.SNESDestroy(snes)

See PETSc manual.

PETSC.SNESGetIterationNumber — Method

PETSC.SNESGetIterationNumber(snes, iter)

See PETSc manual.

PETSC.SNESGetKSP — Method

PETSC.SNESGetKSP(snes, ksp)

See PETSc manual.

PETSC.SNESGetLinearSolveFailures — Method

PETSC.SNESGetLinearSolveFailures(snes, nfails)

See PETSc manual.

PETSC.SNESGetLinearSolveIterations — Method

PETSC.SNESGetLinearSolveIterations(snes, iter)

See PETSc manual.

PETSC.SNESGetNumberFunctionEvals — Method

PETSC.SNESGetNumberFunctionEvals(snes, nfuncs)

See PETSc manual.

PETSC.SNESSetCountersReset — Method

PETSC.SNESSetCountersReset(snes, reset)

See PETSc manual.

PETSC.SNESSetFromOptions — Method

PETSC.SNESSetFromOptions(snes)

See PETSc manual.

PETSC.SNESSetFunction — Method

PETSC.SNESSetFunction(snes, vec, fptr, ctx)

See PETSc manual.

PETSC.SNESSetJacobian — Method

PETSC.SNESSetJacobian(snes, A, P, jacptr, ctx)

See PETSc manual.

PETSC.SNESSetType — Method

PETSC.SNESSetType(snes, type)

See PETSc manual.

PETSC.SNESSolve — Method

PETSC.SNESSolve(snes, b, x)

See PETSc manual.

PETSC.SNESView — Method

PETSC.SNESView(snes, viewer)

See PETSc manual.

PETSC.VecAXPBY — Method

PETSC.VecAXPBY(y, alpha, beta, x)

See PETSc manual.

PETSC.VecAXPY — Method

PETSC.VecAXPY(y, alpha, x)

See PETSc manual.

PETSC.VecAYPX — Method

PETSC.VecAYPX(y, beta, x)

See PETSc manual.

PETSC.VecAssemblyBegin — Method

PETSC.VecAssemblyBegin(vec)

See PETSc manual.

PETSC.VecAssemblyEnd — Method

PETSC.VecAssemblyEnd(vec)

See PETSc manual.

PETSC.VecCopy — Method

PETSC.VecCopy(x, y)

See PETSc manual.

PETSC.VecCreateGhost — Method

PETSC.VecCreateGhost(comm, n, N, nghost, ghosts, vv)

See PETSc manual.

PETSC.VecCreateGhostWithArray — Method

PETSC.VecCreateGhostWithArray(comm, n, N, nghost, ghosts, array, vv)

See PETSc manual.

PETSC.VecCreateMPI — Method

PETSC.VecCreateMPI(comm, n, N, v)

See PETSc manual.

PETSC.VecCreateMPIWithArray — Method

PETSC.VecCreateMPIWithArray(comm, bs, n, N, array, vec)

See PETSc manual.

PETSC.VecCreateSeq — Method

PETSC.VecCreateSeq(comm, n, vec)

See PETSc manual.

PETSC.VecCreateSeqWithArray — Method

PETSC.VecCreateSeqWithArray(comm, bs, n, array, vec)

See PETSc manual.

PETSC.VecDestroy — Method

PETSC.VecDestroy(vec)

See PETSc manual.

PETSC.VecDuplicate — Method

PETSC.VecDuplicate(v, newv)

See PETSc manual.

PETSC.VecGetArray — Method

PETSC.VecGetArray(x, a)

See PETSc manual.

PETSC.VecGetArrayRead — Method

PETSC.VecGetArrayRead(x, a)

See PETSc manual.

PETSC.VecGetArrayWrite — Method

PETSC.VecGetArrayWrite(x, a)

See PETSc manual.

PETSC.VecGetLocalSize — Method

PETSC.VecGetLocalSize(vec, n)

See PETSc manual.

PETSC.VecGetSize — Method

PETSC.VecGetSize(vec, n)

See PETSc manual.

PETSC.VecGetValues — Method

PETSC.VecGetValues(x, ni, ix, y)

See PETSc manual.

PETSC.VecGhostGetLocalForm — Method

PETSC.VecGhostGetLocalForm(g, l)

See PETSc manual.

PETSC.VecGhostRestoreLocalForm — Method

PETSC.VecGhostRestoreLocalForm(g, l)

See PETSc manual.

PETSC.VecNorm — Method

PETSC.VecNorm(x, typ, val)

See PETSc manual.

PETSC.VecPlaceArray — Method

PETSC.VecPlaceArray(vec, array)

See PETSc manual.

PETSC.VecResetArray — Method

PETSC.VecResetArray(vec)

See PETSc manual.

PETSC.VecRestoreArray — Method

PETSC.VecRestoreArray(x, a)

See PETSc manual.

PETSC.VecRestoreArrayRead — Method

PETSC.VecRestoreArrayRead(x, a)

See PETSc manual.

PETSC.VecRestoreArrayWrite — Method

PETSC.VecRestoreArrayWrite(x, a)

See PETSc manual.

PETSC.VecScale — Method

PETSC.VecScale(x, alpha)

See PETSc manual.

PETSC.VecSet — Method

PETSC.VecSet(x, alpha)

See PETSc manual.

PETSC.VecSetOption — Method

PETSC.VecSetOption(x, op, flg)

See PETSc manual.

PETSC.VecSetValues — Method

PETSC.VecSetValues(x, ni, ix, y, iora)

See PETSc manual.

PETSC.VecView — Method

PETSC.VecView(vec, viewer)

See PETSc manual.

PETSC.@PETSC_VIEWER_DRAW_SELF — Macro

@PETSC_VIEWER_DRAW_SELF

See PETSc manual.

PETSC.@PETSC_VIEWER_DRAW_WORLD — Macro

@PETSC_VIEWER_DRAW_WORLD

See PETSc manual.

PETSC.@PETSC_VIEWER_STDOUT_SELF — Macro

@PETSC_VIEWER_STDOUT_SELF

See PETSc manual.

PETSC.@PETSC_VIEWER_STDOUT_WORLD — Macro

@PETSC_VIEWER_STDOUT_WORLD

See PETSc manual.

PETSC.@check_error_code — Macro

@check_error_code expr

Check if expr returns an error code equal to zero(PetscErrorCode). If not, throw an instance of PetscError.