Advanced API

PetscCall.PETSC_DECIDE — Constant

Julia constant storing the PETSC_DECIDE value.

See PETSc manual.

PetscCall.PETSC_DEFAULT — Constant

Julia constant storing the PETSC_DEFAULT value.

See PETSc manual.

PetscCall.PETSC_DETERMINE — Constant

Julia constant storing the PETSC_DETERMINE value.

See PETSc manual.

PetscCall.InsertMode — Type

Julia alias for the InsertMode C enum.

See PETSc manual.

PetscCall.KSP — Type

Julia alias for the KSP C type.

See PETSc manual.

PetscCall.KSPType — Type

Julia alias for KSPType C type.

See PETSc manual.

PetscCall.Mat — Type

Julia alias for the Mat C type.

See PETSc manual.

PetscCall.MatAssemblyType — Type

Julia alias for the MatAssemblyType C enum.

See PETSc manual.

PetscCall.MatDuplicateOption — Type

Julia alias for the MatDuplicateOption C enum.

See PETSc manual.

PetscCall.MatInfo — Type

Julia alias for the MatInfo C struct.

See PETSc manual.

PetscCall.MatInfoType — Type

Julia alias for the MatInfoType C enum.

See PETSc manual.

PetscCall.MatNullSpace — Type

Julia alias for the MatNullSpace C type.

See PETSc manual.

PetscCall.MatReuse — Type

Julia alias for the MatReuse C enum.

See PETSc manual.

PetscCall.MatStructure — Type

Julia alias for the MatStructure C enum.

See PETSc manual.

PetscCall.MatType — Type

Julia alias for MatType C type.

See PETSc manual.

PetscCall.NormType — Type

Julia alias for the NormType C enum.

See PETSc manual.

PetscCall.PC — Type

Julia alias for the PC C type.

See PETSc manual.

PetscCall.PCType — Type

Julia alias for PCType C type.

See PETSc manual.

PetscCall.PetscBool — Type

Julia alias to PetscBool C enum.

See PETSc manual.

PetscCall.PetscDataType — Type

Julia alias to PetscDataType C enum.

See PETSc manual.

PetscCall.PetscError — Type

struct PetscError <: Exception
  code::PetscErrorCode
end

Custom Exception thrown by @check_error_code.

PetscCall.PetscErrorCode — Type

Julia alias to PetscErrorCode C type.

See PETSc manual.

PetscCall.PetscLogDouble — Type

Julia alias to PetscLogDouble C type.

See PETSc manual.

PetscCall.PetscViewer — Type

Julia alias for PetscViewer C type.

See PETSc manual.

PetscCall.SNES — Type

Julia alias for the SNES C type.

See PETSc manual.

PetscCall.Vec — Type

Julia alias for the Vec C type.

See PETSc manual.

PetscCall.VecOption — Type

Julia alias for the VecOption C enum.

See PETSc manual.

PetscCall.KSPCreate — Method

PetscCall.KSPCreate(comm, inksp)

See PETSc manual.

PetscCall.KSPDestroy — Method

PetscCall.KSPDestroy(ksp)

See PETSc manual.

PetscCall.KSPGetIterationNumber — Method

PetscCall.KSPGetIterationNumber(ksp, its)

See PETSc manual.

PetscCall.KSPGetPC — Method

PetscCall.KSPGetPC(ksp, pc)

See PETSc manual.

PetscCall.KSPSetFromOptions — Method

PetscCall.KSPSetFromOptions(ksp)

See PETSc manual.

PetscCall.KSPSetInitialGuessNonzero — Method

PetscCall.KSPSetInitialGuessNonzero(ksp, flg)

See PETSc manual.

PetscCall.KSPSetOperators — Method

PetscCall.KSPSetOperators(ksp, Amat, Pmat)

See PETSc manual.

PetscCall.KSPSetOptionsPrefix — Method

PetscCall.KSPSetOptionsPrefix(ksp, prefix)

See PETSc manual.

PetscCall.KSPSetTolerances — Method

PetscCall.KSPSetTolerances(ksp, rtol, abstol, dtol, maxits)

See PETSc manual.

PetscCall.KSPSetType — Method

PetscCall.KSPSetType(ksp, typ)

See PETSc manual.

PetscCall.KSPSetUp — Method

PetscCall.KSPSetUp(ksp)

See PETSc manual.

PetscCall.KSPSolve — Method

PetscCall.KSPSolve(ksp, b, x)

See PETSc manual.

PetscCall.KSPSolveTranspose — Method

PetscCall.KSPSolveTranspose(ksp, b, x)

See PETSc manual.

PetscCall.KSPView — Method

PetscCall.KSPView(ksp, viewer)

See PETSc manual.

PetscCall.MatAssemblyBegin — Method

PetscCall.MatAssemblyBegin(mat, typ)

See PETSc manual.

PetscCall.MatAssemblyEnd — Method

PetscCall.MatAssemblyEnd(mat, typ)

See PETSc manual.

PetscCall.MatConvert — Method

PetscCall.MatConvert(mat, newtype, reuse, M)

See PETSc manual.

PetscCall.MatCopy — Method

PetscCall.MatCopy(A, B, str)

See PETSc manual.

PetscCall.MatCreate — Method

PetscCall.MatCreate(comm, mat)

See PETSc manual.

PetscCall.MatCreateAIJ — Method

PetscCall.MatCreateAIJ(comm, m, n, M, N, d_nz, d_nnz, o_nz, o_nnz, mat)

See PETSc manual.

PetscCall.MatCreateMPIAIJWithArrays — Method

PetscCall.MatCreateMPIAIJWithArrays(comm, m, n, M, N, i, j, a, mat)

See PETSc manual.

PetscCall.MatCreateMPIAIJWithSplitArrays — Method

PetscCall.MatCreateMPIAIJWithSplitArrays(comm, m, n, M, N, i, j, a, oi, oj, oa, mat)

See PETSc manual.

PetscCall.MatCreateSeqAIJ — Method

PetscCall.MatCreateSeqAIJ(comm, m, n, nz, nnz, mat)

See PETSc manual.

PetscCall.MatCreateSeqAIJWithArrays — Method

PetscCall.MatCreateSeqAIJWithArrays(comm, m, n, i, j, a, mat)

See PETSc manual.

PetscCall.MatDestroy — Method

PetscCall.MatDestroy(A)

See PETSc manual.

PetscCall.MatEqual — Method

PetscCall.MatEqual(A, B, flg)

See PETSc manual.

PetscCall.MatGetInfo — Method

PetscCall.MatGetInfo(mat, flag, info)

See PETSc manual.

PetscCall.MatGetSize — Method

PetscCall.MatGetSize(mat, m, n)

See PETSc manual.

PetscCall.MatGetValues — Method

PetscCall.MatGetValues(mat, m, idxm, n, idxn, v)

See PETSc manual.

PetscCall.MatMPIAIJSetPreallocation — Method

PetscCall.MatMPIAIJSetPreallocation(B, d_nz, d_nnz, o_nz, o_nnz)

See PETSc manual.

PetscCall.MatMult — Method

PetscCall.MatMult(mat, x, y)

See PETSc manual.

PetscCall.MatMultAdd — Method

PetscCall.MatMultAdd(mat, v1, v2, v3)

See PETSc manual.

PetscCall.MatMumpsSetCntl — Method

PetscCall.MatMumpsSetCntl(mat, icntl, val)

See PETSc manual.

PetscCall.MatMumpsSetIcntl — Method

PetscCall.MatMumpsSetIcntl(mat, icntl, val)

See PETSc manual.

PetscCall.MatNullSpaceCreate — Method

PetscCall.MatNullSpaceCreate(comm, has_cnst, n, vecs, sp)

See PETSc manual.

PetscCall.MatNullSpaceCreateRigidBody — Method

PetscCall.MatNullSpaceCreateRigidBody(coords, sp)

See PETSc manual.

PetscCall.MatNullSpaceDestroy — Method

PetscCall.MatNullSpaceDestroy(ns)

See PETSc manual.

PetscCall.MatScale — Method

PetscCall.MatScale(mat, alpha)

See PETSc manual.

PetscCall.MatSetBlockSize — Method

PetscCall.MatSetBlockSize(mat, bs)

See PETSc manual.

PetscCall.MatSetNearNullSpace — Method

PetscCall.MatSetNearNullSpace(mat, nullsp)

See PETSc manual.

PetscCall.MatSetPreallocationCOO — Method

PetscCall.MatSetPreallocationCOO(A, ncoo, coo_i, coo_j)

See PETSc manual.

PetscCall.MatSetSizes — Method

PetscCall.MatSetSizes(A, m, n, M, N)

See PETSc manual.

PetscCall.MatSetType — Method

PetscCall.MatSetType(mat, matype)

See PETSc manual.

PetscCall.MatSetValues — Method

PetscCall.MatSetValues(mat, m, idxm, n, idxn, v, addv)

See PETSc manual.

PetscCall.MatSetValuesCOO — Method

PetscCall.MatSetValuesCOO(A, coo_v, imode)

See PETSc manual.

PetscCall.MatView — Method

PetscCall.MatView(mat, viewer)

See PETSc manual.

PetscCall.MatZeroEntries — Method

PetscCall.MatZeroEntries(mat)

See PETSc manual.

PetscCall.PCFactorGetMatrix — Method

PetscCall.PCFactorGetMatrix(ksp, mat)

See PETSc manual.

PetscCall.PCFactorSetMatSolverType — Method

PetscCall.PCFactorSetMatSolverType(pc, typ)

See PETSc manual.

PetscCall.PCFactorSetUpMatSolverType — Method

PetscCall.PCFactorSetUpMatSolverType(pc)

See PETSc manual.

PetscCall.PCSetType — Method

PetscCall.PCSetType(pc, typ)

See PETSc manual.

PetscCall.PCView — Method

PetscCall.PCView(pc, viewer)

See PETSc manual.

PetscCall.PETSC_VIEWER_DRAW_ — Method

PetscCall.PETSC_VIEWER_DRAW_(comm)

See PETSc manual.

PetscCall.PETSC_VIEWER_STDOUT_ — Method

PetscCall.PETSC_VIEWER_STDOUT_(comm)

See PETSc manual.

PetscCall.PetscDataTypeFromString — Method

PetscDataTypeFromString(name,ptype,found)

See PETSc manual.

PetscCall.PetscDataTypeGetSize — Method

PetscDataTypeGetSize(ptype,size)

See PETSc manual.

PetscCall.PetscFinalize — Method

PetscCall.PetscFinalize()

See PETSc manual.

PetscCall.PetscFinalized — Method

PetscCall.PetscFinalized(flag)

See PETSc manual.

PetscCall.PetscInitializeNoArguments — Method

PetscCall.PetscInitializeNoArguments()

See PETSc manual.

PetscCall.PetscInitializeNoPointers — Method

PetscCall.PetscInitializeNoPointers(argc, args, filename, help)

PetscCall.PetscInitialized — Method

PetscCall.PetscInitialized(flag)

See PETSc manual.

PetscCall.PetscObjectRegisterDestroy — Method

PetscCall.PetscObjectRegisterDestroy(obj)

See PETSc manual.

PetscCall.PetscObjectRegisterDestroyAll — Method

PetscCall.PetscObjectRegisterDestroyAll()

See PETSc manual.

PetscCall.SNESCreate — Method

PetscCall.SNESCreate(comm, snes)

See PETSc manual.

PetscCall.SNESDestroy — Method

PetscCall.SNESDestroy(snes)

See PETSc manual.

PetscCall.SNESGetIterationNumber — Method

PetscCall.SNESGetIterationNumber(snes, iter)

See PETSc manual.

PetscCall.SNESGetKSP — Method

PetscCall.SNESGetKSP(snes, ksp)

See PETSc manual.

PetscCall.SNESGetLinearSolveFailures — Method

PetscCall.SNESGetLinearSolveFailures(snes, nfails)

See PETSc manual.

PetscCall.SNESGetLinearSolveIterations — Method

PetscCall.SNESGetLinearSolveIterations(snes, iter)

See PETSc manual.

PetscCall.SNESGetNumberFunctionEvals — Method

PetscCall.SNESGetNumberFunctionEvals(snes, nfuncs)

See PETSc manual.

PetscCall.SNESSetCountersReset — Method

PetscCall.SNESSetCountersReset(snes, reset)

See PETSc manual.

PetscCall.SNESSetFromOptions — Method

PetscCall.SNESSetFromOptions(snes)

See PETSc manual.

PetscCall.SNESSetFunction — Method

PetscCall.SNESSetFunction(snes, vec, fptr, ctx)

See PETSc manual.

PetscCall.SNESSetJacobian — Method

PetscCall.SNESSetJacobian(snes, A, P, jacptr, ctx)

See PETSc manual.

PetscCall.SNESSetType — Method

PetscCall.SNESSetType(snes, type)

See PETSc manual.

PetscCall.SNESSolve — Method

PetscCall.SNESSolve(snes, b, x)

See PETSc manual.

PetscCall.SNESView — Method

PetscCall.SNESView(snes, viewer)

See PETSc manual.

PetscCall.VecAXPBY — Method

PetscCall.VecAXPBY(y, alpha, beta, x)

See PETSc manual.

PetscCall.VecAXPY — Method

PetscCall.VecAXPY(y, alpha, x)

See PETSc manual.

PetscCall.VecAYPX — Method

PetscCall.VecAYPX(y, beta, x)

See PETSc manual.

PetscCall.VecAssemblyBegin — Method

PetscCall.VecAssemblyBegin(vec)

See PETSc manual.

PetscCall.VecAssemblyEnd — Method

PetscCall.VecAssemblyEnd(vec)

See PETSc manual.

PetscCall.VecCopy — Method

PetscCall.VecCopy(x, y)

See PETSc manual.

PetscCall.VecCreateGhost — Method

PetscCall.VecCreateGhost(comm, n, N, nghost, ghosts, vv)

See PETSc manual.

PetscCall.VecCreateGhostWithArray — Method

PetscCall.VecCreateGhostWithArray(comm, n, N, nghost, ghosts, array, vv)

See PETSc manual.

PetscCall.VecCreateMPI — Method

PetscCall.VecCreateMPI(comm, n, N, v)

See PETSc manual.

PetscCall.VecCreateMPIWithArray — Method

PetscCall.VecCreateMPIWithArray(comm, bs, n, N, array, vec)

See PETSc manual.

PetscCall.VecCreateSeq — Method

PetscCall.VecCreateSeq(comm, n, vec)

See PETSc manual.

PetscCall.VecCreateSeqWithArray — Method

PetscCall.VecCreateSeqWithArray(comm, bs, n, array, vec)

See PETSc manual.

PetscCall.VecDestroy — Method

PetscCall.VecDestroy(vec)

See PETSc manual.

PetscCall.VecDuplicate — Method

PetscCall.VecDuplicate(v, newv)

See PETSc manual.

PetscCall.VecGetArray — Method

PetscCall.VecGetArray(x, a)

See PETSc manual.

PetscCall.VecGetArrayRead — Method

PetscCall.VecGetArrayRead(x, a)

See PETSc manual.

PetscCall.VecGetArrayWrite — Method

PetscCall.VecGetArrayWrite(x, a)

See PETSc manual.

PetscCall.VecGetLocalSize — Method

PetscCall.VecGetLocalSize(vec, n)

See PETSc manual.

PetscCall.VecGetSize — Method

PetscCall.VecGetSize(vec, n)

See PETSc manual.

PetscCall.VecGetValues — Method

PetscCall.VecGetValues(x, ni, ix, y)

See PETSc manual.

PetscCall.VecGhostGetLocalForm — Method

PetscCall.VecGhostGetLocalForm(g, l)

See PETSc manual.

PetscCall.VecGhostRestoreLocalForm — Method

PetscCall.VecGhostRestoreLocalForm(g, l)

See PETSc manual.

PetscCall.VecNorm — Method

PetscCall.VecNorm(x, typ, val)

See PETSc manual.

PetscCall.VecPlaceArray — Method

PetscCall.VecPlaceArray(vec, array)

See PETSc manual.

PetscCall.VecResetArray — Method

PetscCall.VecResetArray(vec)

See PETSc manual.

PetscCall.VecRestoreArray — Method

PetscCall.VecRestoreArray(x, a)

See PETSc manual.

PetscCall.VecRestoreArrayRead — Method

PetscCall.VecRestoreArrayRead(x, a)

See PETSc manual.

PetscCall.VecRestoreArrayWrite — Method

PetscCall.VecRestoreArrayWrite(x, a)

See PETSc manual.

PetscCall.VecScale — Method

PetscCall.VecScale(x, alpha)

See PETSc manual.

PetscCall.VecSet — Method

PetscCall.VecSet(x, alpha)

See PETSc manual.

PetscCall.VecSetOption — Method

PetscCall.VecSetOption(x, op, flg)

See PETSc manual.

PetscCall.VecSetValues — Method

PetscCall.VecSetValues(x, ni, ix, y, iora)

See PETSc manual.

PetscCall.VecView — Method

PetscCall.VecView(vec, viewer)

See PETSc manual.

PetscCall.VecZeroEntries — Method

PetscCall.VecZeroEntries(v)

See PETSc manual.

PetscCall.@PETSC_VIEWER_DRAW_SELF — Macro

@PETSC_VIEWER_DRAW_SELF

See PETSc manual.

PetscCall.@PETSC_VIEWER_DRAW_WORLD — Macro

@PETSC_VIEWER_DRAW_WORLD

See PETSc manual.

PetscCall.@PETSC_VIEWER_STDOUT_SELF — Macro

@PETSC_VIEWER_STDOUT_SELF

See PETSc manual.

PetscCall.@PETSC_VIEWER_STDOUT_WORLD — Macro

@PETSC_VIEWER_STDOUT_WORLD

See PETSc manual.

PetscCall.@check_error_code — Macro

@check_error_code expr

Check if expr returns an error code equal to zero(PetscErrorCode). If not, throw an instance of PetscError.